fix(agent): run the browser agent in the chat drawer too, not just the pill
Phase 3 flipped only the FloatingAssistant composer pill to the browser engine
on synced boards; the CopilotSheet drawer (and its `<Chat>`) still mounted the
backend agent with local=false. Opening the drawer hides the pill, so a user who
chatted via the drawer hit the server agent — and turns sent from the pill (in
the local store) never appeared in the drawer, which reads the backend chat.

Thread `local`/`syncTranscript` from BoardView into CopilotSheet -> Chat so both
surfaces run the same engine, and move the local-store init (seed + openBoard)
up to BoardView — it stays mounted while either surface shows, so the pill and
drawer share one conversation (they never mount together). FloatingAssistant now
just consumes the store. Flag off -> local=false everywhere, unchanged.

The standalone /chats/$id page is still backend-only (not board-scoped) — a
separate follow-up.
